==> 01_CombineData.ipynb <==
import anndata as ad
import scvelo as scv
import scanpy as sc
import pandas as pd
import numpy as np

## Read the files

d1mtx = ad.read_mtx('D1-D3-D5-byHaiping/D1/filtered_feature_bc_matrix/matrix.mtx')
d1barcodes = pd.read_csv('D1-D3-D5-byHaiping/D1/filtered_feature_bc_matrix/barcodes.tsv',header=None)
d1features = pd.read_csv('D1-D3-D5-byHaiping/D1/filtered_feature_bc_matrix/features.tsv',header=None,sep='\t')

d3mtx = ad.read_mtx('D1-D3-D5-byHaiping/D3/filtered_feature_bc_matrix/matrix.mtx')
d3barcodes = pd.read_csv('D1-D3-D5-byHaiping/D3/filtered_feature_bc_matrix/barcodes.tsv',header=None)
d3features = pd.read_csv('D1-D3-D5-byHaiping/D3/filtered_feature_bc_matrix/features.tsv',header=None,sep='\t')

d5mtx = ad.read_mtx('D1-D3-D5-byHaiping/D5/filtered_feature_bc_matrix/matrix.mtx')
d5barcodes = pd.read_csv('D1-D3-D5-byHaiping/D5/filtered_feature_bc_matrix/barcodes.tsv',header=None)
d5features = pd.read_csv('D1-D3-D5-byHaiping/D5/filtered_feature_bc_matrix/features.tsv',header=None,sep='\t')

d1mtx = d1mtx.transpose()
d3mtx = d3mtx.transpose()
d5mtx = d5mtx.transpose()

## Assign cell and gene lables

def cleanBarcodes(df):
    barcodes = df.iloc[:,0].values
    return [barcode[:-2] for barcode in barcodes]

d1barcodes = cleanBarcodes(d1barcodes)
d3barcodes = cleanBarcodes(d3barcodes)
d5barcodes = cleanBarcodes(d5barcodes)

d1mtx.obs['cellId'] = d1barcodes
d3mtx.obs['cellId'] = d3barcodes
d5mtx.obs['cellId'] = d5barcodes

def assignVarValues(adata,df):
    adata.var['geneName'] = df.iloc[:,1].values
    adata.var['geneId'] = df.iloc[:,0].values

assignVarValues(d1mtx,d1features)
assignVarValues(d3mtx,d3features)
assignVarValues(d5mtx,d5features)

d1lables = pd.read_csv('finalSampleLabelsD1.txt',sep='\t')
d3lables = pd.read_csv('finalSampleLabelsD3.txt',sep='\t')
d5lables = pd.read_csv('finalSampleLabelsD5.txt',sep='\t')

d1lables.head()

def removeDuplicates(df):
    dup = df.duplicated('cellBarcodes',keep='last')
    notDup = [not x for x in dup]
    df = df.loc[notDup,:].sort_values('cellBarcodes')
    return df

d1lables = removeDuplicates(d1lables)
d3lables = removeDuplicates(d3lables)
d5lables = removeDuplicates(d5lables)

d1mtx.obs = d1mtx.obs.set_index('cellId')
d3mtx.obs = d3mtx.obs.set_index('cellId')
d5mtx.obs = d5mtx.obs.set_index('cellId')

d1mtx = ad.AnnData(d1mtx[d1mtx.obs.index.isin(d1lables['cellBarcodes'].values)])
d3mtx = ad.AnnData(d3mtx[d3mtx.obs.index.isin(d3lables['cellBarcodes'].values)])
d5mtx = ad.AnnData(d5mtx[d5mtx.obs.index.isin(d5lables['cellBarcodes'].values)])

d1lables.set_index('cellBarcodes',inplace=True)
d3lables.set_index('cellBarcodes',inplace=True)
d5lables.set_index('cellBarcodes',inplace=True)

d1mtx.shape

d1mtx.obs.head()

d1mtx.obs['cellLabel'] = d1lables.loc[list(d1mtx.obs.index.values),'final.calls'].values
d3mtx.obs['cellLabel'] = d3lables.loc[list(d3mtx.obs.index.values),'final.calls'].values
d5mtx.obs['cellLabel'] = d5lables.loc[list(d5mtx.obs.index.values),'final.calls'].values

b1 = list(d1barcodes)
b3 = list(d3barcodes)
b5 = list(d5barcodes)

# common cell barcodes
print(len([x for x in b1 if x in b3]))
print(len([x for x in b1 if x in b5]))
print(len([x for x in b5 if x in b3]))

adata = d1mtx.concatenate(d3mtx,join='outer')
adata = adata.concatenate(d5mtx,join='outer')

adata

adata.var = adata.var.drop(['geneName-0-0', 'geneId-0-0', 'geneName-1-0', 'geneId-1-0'],axis=1)

adata.var = adata.var.rename({'geneName-1':'geneName','geneId-1':'geneId'},axis=1)

adata.write_h5ad('mergedAdata.h5ad')
