==> 12_cogapsVisuals_collagen.Rmd <==
---
title: "CoGAPS Visuals"
author: "Melanie"
date: "7/12/2021"
output: html_document
---



```{r libraries , message = FALSE, warning = FALSE}
library(CoGAPS)
library('ComplexHeatmap')
```



```{r loadData}
## Load the CoGAPS result
gwCogapsRes <- readRDS("gwCogapsRes/emtGwCogaps_20.rds")
## Load the EMT genes
emtGenes <- read.csv("geneList.csv", header = FALSE)
```


```{r patternMarkerRank}
## Calculate the pattern marker statistics ----
## Duplicate the result to manipulate 
filteredCogaps <- gwCogapsRes

## Eliminate the cell line patterns
filteredCogaps@featureLoadings <- filteredCogaps@featureLoadings[, !colnames(filteredCogaps@featureLoadings) %in%c("Pattern_2", "Pattern_13")]

## Run pattern marker function
patternMarkerResults <- patternMarkers(filteredCogaps)

## Extract the rank results for the EMT genes
emtGeneRanks <- patternMarkerResults$PatternMarkerRanks[rownames(patternMarkerResults$PatternMarkerRanks) %in% emtGenes$V1, ]

## Save the csv
write.csv(emtGeneRanks, file = "emtGeneRanks.csv")
```


```{r,amplitudeHeatmap}
## Heatmap of the amplitude matrix ----
## Extract the amplitude matrix while dropping the cell line patterns
amplitude <- gwCogapsRes@featureLoadings[rownames(gwCogapsRes@featureLoadings) %in% emtGenes$V1, 
    !colnames(gwCogapsRes@featureLoadings) %in%c("Pattern_2", "Pattern_13")]

## Scale and center the data
scaledmat<- t(apply(amplitude, 1, scale))

## Reassign the column names
colnames(scaledmat) <- colnames(amplitude)


## Create the heatmap
Heatmap(scaledmat, name = "Amplitude",cluster_columns = TRUE,
        show_row_dend = FALSE, 
        show_column_dend = FALSE,
        column_title = "Amplitude Heatmap of EMT Genes (nPattern = 20)")


```
