==> 13_Figure.Rmd <==
---
title: "Figure"
author: "Melanie and Joe"
date: "7/13/2021"
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
```

```{r libraries}
library(monocle3)
library(CoGAPS)
library(ggplot2)
library(dplyr)
library('ComplexHeatmap')
library(forcats)
library(knitr)
sessionInfo()
```

```{r loadData}
## Load the cds objects
load("Results/cellCycle_EJFPsuedotimeResults2Mar2020.Rda")
## Load the CoGAPS 20 Pattern Result
cogapsRes <- readRDS("Results/gwCogapsRes/emtGwCogaps_20.rds")
## Load the pathways results for the emt pathwyay
emtPatterns <- read.csv(file = "Results/msigdb/emtPatterns.csv", header = TRUE)
## Load the pattern markers
patternMarkerResults <- read.csv(file = "Results/patternMarkers_emtGwCogaps_20.csv")
## Load the EMT genes
emtGenes <- read.csv("Results/geneList.csv", header = FALSE)
```
## Panel A : Cell cycle
```{r}
## Plot the cell cycle
plot_cells(cdsNorm, color_cells_by = "phase", label_cell_groups = FALSE) +
    ggtitle("Cell Cycle Phase")
```

## Panel B: Day
```{r}
## Plot the cell cycle
plot_cells(cdsNorm, color_cells_by = "day", label_cell_groups = FALSE) +
    ggtitle("Day")
```

## Panel C : Violin plots of Cdh1 / Vim by cell cycle state
```{r violinPlot}
## Violin plot
## Subset to Vim and CDH1
cdsSubset <- cdsNorm[rowData(cdsNorm)$gene_short_name %in% c("Vim", "Cdh1"), ]
plot_genes_violin(cdsSubset, group_cells_by = "phase", ncol = 2)
```

## Panel D : Barplot of -10*log10(p values) for the hallmark EMT pathway
```{r}
## Make barplot for the hallmark emt pathway----
## Read in the emt pathway tables
emtPatterns <- read.csv("Results/msigdb/HighPValueResults/emtResults.csv")
## Add the column with the updated p value
emtPatterns$'-10*log10(adj.p.value)' <- (-10)*log10(emtPatterns$FDR.q.value)

## Make the barplot

emtPatterns %>% 
    ## Reorders the barplots
    mutate(Pattern=fct_reorder(Pattern, desc(p.value) )) %>%
    ggplot(aes_string(y = "-10*log10(p.value)", x = "Pattern", fill = "Pattern")) +
    ## Specifies barplot
    geom_col() +
    ## Flips the coordinates
    coord_flip() +
    ## Makes the background white
    theme_minimal() +
    ## Add title
    ggtitle("Hallmark EMT Pathway") +
    ## This creates the dotted line at .05 value 
    geom_hline(yintercept=c(13.0103), linetype="dotted") + # Add veritcle line to show significances
    ## Adds the q values
    geom_text(aes(label=FDR.q.value), hjust = -.04) +
    ## Removes legend
    theme(legend.position = "none") +
    ## specifies limits 
    ylim(0, 450)

```


## Panel E : Heatmap with * added to the heatmap if it's a pattern marker gene 
```{r, fig.height=6, fig.width=8}
## Heatmap of the amplitude matrix ----
## Extract the amplitude matrix while dropping the cell line patterns
amplitude <- cogapsRes@featureLoadings[rownames(cogapsRes@featureLoadings) %in% emtGenes$V1, 
    colnames(cogapsRes@featureLoadings) %in% c("Pattern_5","Pattern_8","Pattern_18", "Pattern_14")]
## Reorder the scaledmat
amplitude <- amplitude[, c("Pattern_5", "Pattern_8", "Pattern_18", "Pattern_14")] 
## Scale and center the data
scaledmat<- t(apply(amplitude, 1, scale))

## Reassign the column names
colnames(scaledmat) <- colnames(amplitude)

## Add * if the gene is a pattern marker for 
## initalize an empty matrix
small_mat <- matrix("", nrow = nrow(scaledmat), ncol= ncol(scaledmat))
## Add row andcolumn names to matrix
rownames(small_mat) <- rownames(scaledmat)
colnames(small_mat) <- colnames(scaledmat)



## Add * if the gene is pattern marker
for (i in rownames(small_mat)){
    for(j in colnames(small_mat)){
        if( paste0(i) %in% patternMarkerResults[,paste0(j)]) {
            small_mat[paste(i), paste0(j)] <- "*"
        }
    }
}


## Create the heatmap
Heatmap(t(scaledmat), name = "Amplitude",cluster_columns = TRUE,
        show_row_dend = FALSE, 
        show_column_dend = FALSE,
        cluster_rows = FALSE,
        row_names_side = "left",
        column_names_side = "bottom",
        column_names_rot = 45,
        column_names_gp = gpar(fontsize=10),
        column_title = "Amplitude Heatmap of EMT Genes (nPattern = 20)",
        cell_fun = function(j, i, x, y, width, height, fill) {
        grid.text(sprintf(small_mat[j, i]), x, y, gp = gpar(fontsize = 10))
})


```


## Panel F : UMAPS colored by pattern weights for patterns 5, 8, 14, and 18 
```{r}
## Create the pattern umaps ----
cdsNorm@colData <- cbind(colData(cdsNorm), cogapsRes@sampleFactors[colnames(cdsNorm), ])

## Pattern 5
plot_cells(cdsNorm, color_cells_by = "Pattern_5", label_cell_groups = FALSE)

## Pattern 8
plot_cells(cdsNorm, color_cells_by = "Pattern_8", label_cell_groups = FALSE)

## Pattern 12
plot_cells(cdsNorm, color_cells_by = "Pattern_14", label_cell_groups = FALSE) 

## Pattern 14
plot_cells(cdsNorm, color_cells_by = "Pattern_18", label_cell_groups = FALSE)
```


## Panel G : Show percentage of cells in each pattern
```{r}
## Subset the cogaps results to the specific patterns
## Duplicate the result to manipulate 
filteredCogaps <- cogapsRes
## Eliminate the cell line patterns
filteredCogaps@sampleFactors <- filteredCogaps@sampleFactors[, !colnames(filteredCogaps@sampleFactors) %in%c("Pattern_2", "Pattern_13")]
## Run the pattern marker statistic on the cells 
cellPatternMarker <- patternMarkers(filteredCogaps, axis = 2)
## Add the pattern to the cell marker result
for(i in 1:length(cellPatternMarker$PatternMarkers)){
    ## Concert to a df
    cellPatternMarker$PatternMarkers[[i]] <- as.data.frame(cellPatternMarker$PatternMarkers[[i]])
    ## create column name
    colnames(cellPatternMarker$PatternMarkers[[i]]) <- "CellID"
    cellPatternMarker$PatternMarkers[[i]]$Pattern <- names(cellPatternMarker$PatternMarkers[i])
}
## Combined the tables
cellPatternMarkerDataFrame <- do.call("rbind", cellPatternMarker$PatternMarkers)
## Update rownames
rownames(cellPatternMarkerDataFrame) <- cellPatternMarkerDataFrame$CellID

## Add the pattern assignment to the cds
colData(cdsNorm)$AssignedPattern <- cellPatternMarkerDataFrame[colnames(cdsNorm), "Pattern"]
## Plot the data
plot_cells(cdsNorm, color_cells_by = "AssignedPattern", label_cell_groups = FALSE)
## Show pattern 5, 8, and 12 only
plot_cells(cdsNorm, color_cells_by = "AssignedPattern", label_cell_groups = FALSE) +
    scale_color_manual(values = c("gray80","gray80","gray80","black","#2297E6",
                                  "gray80","gray80","gray80","gray80","gray80",
                                  "gray80","gray80","#F5C710","gray80","gray80",
                                  "#CD0BBC","gray80"))
# Generate a table which counts cells assigned to each pattern
countsTable <- table(colData(cdsNorm)$AssignedPattern)
## make a proportion table with the counts table
propTable <- as.data.frame(prop.table(countsTable))
## round the proportions
propTable$Freq <- round(propTable$Freq, digits = 4)
## Merge the two tables
combinedTable <- merge(as.data.frame(countsTable), propTable, by = "Var1")
## Rename columns
colnames(combinedTable) <- c("Pattern", "Number of Cells in Pattern", "Proportion of Cells in Pattern")
## Create table
kable(combinedTable)

```

```{r}
## Make umaps for the gene's in gene list
pdf("Results/emtGeneUmaps.pdf")
for (i in 1:length(emtGenes$V1)){
    print(plot_cells(cdsNorm, genes = paste0(emtGenes$V1[i])))
}
dev.off()

```
## Panel H : UMAP Showing Vim and Cdh1 positive/negative cells
```{r}
## Subset the cds p data to Vim and Cdh1
countsDataFrame <- as.data.frame(as.matrix(counts(cdsNorm)))
countsSubset <- t(countsDataFrame[rownames(countsDataFrame) %in% c("Vim", "Cdh1"), ])
## Classify by expression of the genes of interst
classifiedCounts <- as.data.frame(countsSubset) %>%
    mutate(classification = ifelse(Vim > 0 & Cdh1 > 0, "Double Positive",
                                   ifelse(Vim > 0 & Cdh1 == 0, "Vim Positive",
                                   ifelse(Vim == 0 & Cdh1 > 0, "Cdh1 Positive",
                                   "Double Negative"))))
## Add the classification to the cds
colData(cdsNorm)$Classification <- classifiedCounts[colnames(cdsNorm), "classification"]
## Plot the classification
plot_cells(cdsNorm, color_cells_by = "Classification", label_cell_groups = FALSE) + scale_color_manual(values = c("cadetblue2","gray50","purple","orange"))

## Add table for the number of cells in each classification
## Create table
kable(table(colData(cdsNorm)$Classification))
```

## Panel I : Gene Expression Pattern Plots *Collagen Data only*
```{r}
#Assigning the data
## Extract the cogaps pattern weights
pattern.weights <- (cogapsRes@sampleFactors[, colnames(cogapsRes@sampleFactors) %in% c("Pattern_5","Pattern_8","Pattern_18", "Pattern_14")])
## Check the column names, dimensions and rownames
colnames(pattern.weights)
dim(pattern.weights)
head(rownames(pattern.weights))

##Extract day info
day.info <- colData(cdsNormCollagen)[, colnames(colData(cdsNormCollagen)) == "day", drop =FALSE]

##Extract gene expression info
epithelialGenes <- emtGenes$V1[c(1:6,16)]
expression.data <- cdsNormCollagen[rownames(cdsNormCollagen) %in% epithelialGenes, ]
expression.data <- t(exprs(expression.data))
expression.data <- as.data.frame(log10(expression.data))

## Merge expression and day info
mergedDayExpr <- cbind(day.info, expression.data)

## Merge in pattern matrix
data <- as.data.frame(merge(mergedDayExpr, pattern.weights, all.x = TRUE, by = "row.names"))


#Test Plot with 1 gene and 1 pattern
ggplot(data,aes(x=(Pattern_5),y=(Vim), colour=day)) + geom_point() + labs(y="log10(Expression)",title="Vim",x="Pattern 5") + theme(axis.text = element_blank()) + geom_smooth(color='black', se=FALSE) + theme_minimal() + theme(plot.title = element_text(hjust = 0.5))


loop1 <- data[,3:9]


#Generating pdf of Pattern 5
 
  pdf(file = "/finalfigures/P5_Expression_Patterns.pdf")
  
  #Plot List
  plot_list <- list()
  
  for (i in 1:ncol(loop1)){
    
    print(ggplot(data,aes((Pattern_5),(loop1[,i]),colour=day)) + geom_point() + labs(title=colnames(loop1[i]), x="Pattern 5", y="log10(Expression)") + theme(axis.text = element_blank())+ geom_smooth(color='black',se=FALSE) + theme_minimal() + theme(plot.title = element_text(hjust = 0.5)))

  }
  
  dev.off()

  
#Generating pdf of Pattern 8
 
  pdf(file = "/finalfigures/P8_Expression_Patterns.pdf")
  
  #Plot List
  plot_list <- list()
  
  for (i in 1:ncol(loop1)){
    
    print(ggplot(data,aes((Pattern_8),(loop1[,i]),colour=day)) + geom_point() + labs(title=colnames(loop1[i]), x="Pattern 8", y="log10(Expression)") + theme(axis.text = element_blank())+ geom_smooth(color='black',se=FALSE) + theme_minimal() + theme(plot.title = element_text(hjust = 0.5)))
  
  }

  dev.off()
  
  
  #Generating pdf of Pattern 14
 
  pdf(file = "/finalfigures/P14_Expression_Patterns.pdf")
  
  #Plot List
  plot_list <- list()
  
  for (i in 1:ncol(loop1)) {
    
    print(ggplot(data,aes((Pattern_14),(loop1[,i]),colour=day)) + geom_point() + labs(title=colnames(loop1[i]), x="Pattern 14", y="log10(Expression)") + theme(axis.text = element_blank())+ geom_smooth(color='black',se=FALSE) + theme_minimal() + theme(plot.title = element_text(hjust = 0.5)))
  
  }
  
  dev.off()
  
  
  #Generating pdf of Pattern18
 
  pdf(file = "/finalfigures/P18_Expression_Patterns.pdf")
  
  #Plot List
  plot_list <- list()
  
  for (i in 1:ncol(loop1)){
    
    print(ggplot(data,aes((Pattern_18),(loop1[,i]),colour=day)) + geom_point() + labs(title=colnames(loop1[i]), x="Pattern 18", y="log10(Expression)") + theme(axis.text = element_blank())+ geom_smooth(color='black',se=FALSE) + theme_minimal() + theme(plot.title = element_text(hjust = 0.5)))
  
  }

  dev.off()

as.matrix(data)

```

##Panel J: Patterns vs Pseudotime
```{r}
#Assigning the data
## Extract the cogaps pattern weights
pattern.weights <- (cogapsRes@sampleFactors[, colnames(cogapsRes@sampleFactors) %in% c("Pattern_5","Pattern_8","Pattern_18", "Pattern_14")])
## Check the column names, dimensions and rownames
colnames(pattern.weights)
dim(pattern.weights)
head(rownames(pattern.weights))

##Extract day info
day.info <- colData(cdsNormCollagen)[, colnames(colData(cdsNormCollagen)) == "day", drop =FALSE]

##Generating pseudotime info
Pseudotime <- pseudotime(cdsNormCollagen)

## Merge pseudotime and day info
mergedDayPseudo <- cbind(day.info, Pseudotime)

## Merge in pattern matrix
pseudoplots.data <- as.data.frame(merge(mergedDayPseudo, pattern.weights, all.x = TRUE, by = "row.names"))

pdf(file="/FinalFigures/Pattern_v_Pseudotime.pdf")

#Pattern 5
ggplot(pseudoplots.data,aes(x=(Pseudotime),y=(Pattern_5), colour=day)) + geom_point() + labs(y="Pattern Weight",title="Pattern 5",x="Pseudotime") + theme(axis.text = element_blank()) + geom_smooth(color='black', se=FALSE) + theme_minimal() + theme(plot.title = element_text(hjust = 0.5))

#Pattern 8
ggplot(pseudoplots.data,aes(x=(Pseudotime),y=(Pattern_8), colour=day)) + geom_point() + labs(y="Pattern Weight",title="Pattern 8",x="Pseudotime") + theme(axis.text = element_blank()) + geom_smooth(color='black', se=FALSE) + theme_minimal() + theme(plot.title = element_text(hjust = 0.5))

#Pattern 14
ggplot(pseudoplots.data,aes(x=(Pseudotime),y=(Pattern_14), colour=day)) + geom_point() + labs(y="Pattern Weight",title="Pattern 14",x="Pseudotime") + theme(axis.text = element_blank()) + geom_smooth(color='black', se=FALSE) + theme_minimal() + theme(plot.title = element_text(hjust = 0.5))

#Pattern 5
ggplot(pseudoplots.data,aes(x=(Pseudotime),y=(Pattern_18), colour=day)) + geom_point() + labs(y="Pattern Weight",title="Pattern 18",x="Pseudotime") + theme(axis.text = element_blank()) + geom_smooth(color='black', se=FALSE) + theme_minimal() + theme(plot.title = element_text(hjust = 0.5))

dev.off()
```
