==> 15_CoGAPSData_matrigel.Rmd <==
---
title: "EMT single cell matrigel analysis"
author: "Melanie Loth"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

```{r libraries}
library(monocle3)
library(CoGAPS)
sessionInfo()
```

# Load in proceessed data and imputed data

```{r}
load('Results/EJFPsuedotimeResults2Mar2020.Rda')
## Load a Cogaps results from the collagen so the genes can be extracted
cogapsRes <- readRDS("Results/emtGwCogaps_20.rds")

```

# Subset to the same set of genes used in the collagen analysis


# Matriegl data with subset genes for CoGAPS analysis

```{r}
sbstCDSCoGAPS <- cdsNormMatrigel[unique(rownames(cogapsRes@featureLoadings)),]
# save the object
saveRDS(sbstCDSCoGAPS, file = "Results/15_sbstCdsMatrigel.rds")
saveRDS(counts(sbstCDSCoGAPS), file = "Results/15_countsSbstCdsMatrigel.rds")
```
