==> 17_ewald_gwCogapsAnalysis_matrigel.Rmd <==
---
title: "Matrigel Cogaps Analysis"
author: "Melanie and Joe"
date: "8/2/2021"
output: html_document
---

```{r libraries}
library(monocle3)
library(tibble)
library(ggplot2)
library(gridExtra)
library(CoGAPS)
#library('tricycle')
sessionInfo() 
```

```{r loadData}
# CDS object
sbstCDSCoGAPS <- readRDS("/15_sbstCdsMatrigel.rds")
# CoGAPS result paths
gwCogapsPath <- list.files(path = "/gwCogapsRes", full.names = TRUE)
# Read in the results
gwCogapsRes <- lapply(gwCogapsPath, readRDS)
# Provide names for the cogaps results, this will help with plot names
names(gwCogapsRes) <- gsub("gwCogapsRes/|.rds", "", gwCogapsPath)
```



```{r metadataVisuals}
# Plot the cells in CDS to visualize the umap
plot_cells(sbstCDSCoGAPS, color_cells_by = 'batch')

# Make batch categorical
colData(sbstCDSCoGAPS)$batch <- as.factor(colData(sbstCDSCoGAPS)$batch)
colData(sbstCDSCoGAPS)$mouseDay <- colData(sbstCDSCoGAPS)$'mouse-day' 
pdf("metaData_umap.pdf")
# Plot the meta data variables on UMAP----
    # retrieve the metadata variables
    metaData <- colnames(colData(sbstCDSCoGAPS))
    # drop index as a variable
    metaData <- metaData[!metaData %in% c("index")]
    # Loop through the metaData and plot
    for (i in metaData){
        # plot the datta
        print(plot_cells(sbstCDSCoGAPS, color_cells_by = paste0(i), 
              label_cell_groups = FALSE))
    }
dev.off()
```

```{r cogapsVisuals}
for(i in 1:length(gwCogapsRes)){
    # Extract the pattern matrix
    patMat <- gwCogapsRes[[i]]@sampleFactors
    # initalize a temporary cds to manipulate
    tempCds <- sbstCDSCoGAPS
    # Merge with the column data in the cds
    tempCds@colData <- cbind(colData(tempCds), patMat[colnames(tempCds), ])
    # Make UMAP plots for the results
        # initalize the pdf
        pdf(paste0(names(gwCogapsRes[i]), "_umaps.pdf"))
        # initalize the list 
        umap_list <- list()
        for (j in colnames(patMat)){
            # plot the data
            print(plot_cells(tempCds, color_cells_by = paste0(j),
                  label_cell_groups = FALSE,show_trajectory_graph = FALSE))
                  
                # + scale_color_gradient2(name = paste0(paste0(j)), low = "grey96", high = "red"))
        }
        dev.off()
        # Make Box plots for the results
        # Extract the colData for ggplot
        colDataDf <- as.data.frame(tempCds@colData)
        
        # Initalize a pdf
        pdf(paste0(names(gwCogapsRes[i]), "_boxplots.pdf"))
        for(k in colnames(patMat)){
            # initalize the list
            boxplot_list <- list()
            # extract the categorical variables
            categoricalVariables <- c("batch", "cellLabel", "mouse", "media", 
                                      "day", "mouseDay", "mediaday")
            for(l in categoricalVariables){
                boxplot_list[[l]] <-(ggplot(colDataDf, aes_string(x = paste(l), 
                                        y = paste(k))) +
                            geom_boxplot() +
                            theme(axis.text.x = element_text(angle = 45, hjust = 1, size = 4)))
            }
            print(do.call(grid.arrange, boxplot_list))
        }
        dev.off()
}
```
```{r}
# Patttern Correlations----
# Extract and matrices and append the result name
gaps1 <- gwCogapsRes[[1]]@sampleFactors
colnames(gaps1) <- paste0(names(gwCogapsRes[1]), "_", colnames(gaps1))
gaps2 <- gwCogapsRes[[2]]@sampleFactors
colnames(gaps2) <- paste0(names(gwCogapsRes[2]), "_", colnames(gaps2))
# Merge the pattern matricies
mergedGaps <- cbind(gaps1, gaps2[rownames(gaps1), ])
# Calculate cor
corMat <- cor(mergedGaps)
# Plot cor
heatmap(corMat)
```

```{r}
# Pattern marker statistics-----
## Run pattern marker statistic for p = 10

filteredCogaps <- gwCogapsRes[[1]]

patternMarkerResults <- patternMarkers(filteredCogaps)


#make into a data frame for csv
df <- data.frame(lapply(patternMarkerResults$PatternMarkers, "length<-", max(lengths(patternMarkerResults$PatternMarkers))))

write.csv(df, file="patternMarkers_emtGwCogapsMatrigel_10.csv", row.names = FALSE, na="")

```

```{r}
# Pattern marker statistics-----
## Run pattern marker statistic for p = 12

filteredCogaps <- gwCogapsRes[[2]]

patternMarkerResults <- patternMarkers(filteredCogaps)


#make into a data frame for csv
df <- data.frame(lapply(patternMarkerResults$PatternMarkers, "length<-", max(lengths(patternMarkerResults$PatternMarkers))))

write.csv(df, file="patternMarkers_emtGwCogapsMatrigel_12.csv", row.names = FALSE, na="")

```

```{r}
# Pattern marker statistics-----
## Run pattern marker statistic for p = 14

filteredCogaps <- gwCogapsRes[[3]]

patternMarkerResults <- patternMarkers(filteredCogaps)


#make into a data frame for csv
df <- data.frame(lapply(patternMarkerResults$PatternMarkers, "length<-", max(lengths(patternMarkerResults$PatternMarkers))))

write.csv(df, file="patternMarkers_emtGwCogapsMatrigel_14.csv", row.names = FALSE, na="")

```

```{r}
# Pattern marker statistics-----
## Run pattern marker statistic for p = 16

filteredCogaps <- gwCogapsRes[[4]]

patternMarkerResults <- patternMarkers(filteredCogaps)


#make into a data frame for csv
df <- data.frame(lapply(patternMarkerResults$PatternMarkers, "length<-", max(lengths(patternMarkerResults$PatternMarkers))))

write.csv(df, file="patternMarkers_emtGwCogapsMatrigel_16.csv", row.names = FALSE, na="")

```
